--- examples/tool_call_configs/pythonic.json
@@ -1,8 +1,0 @@
-{
-  "format": "Pythonic",
-  "parallel_tool_start_token": null,
-  "parallel_tool_end_token": null,
-  "call_start_token": null,
-  "call_end_token": null,
-  "use_parameters_field": null
-}
