+++ lib/llm/src/postprocessor/tool_calling/mod.rs
@@ -13,7 +13,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub fn try_tool_call_parse_pythonic(message: &str) -> anyhow::Result<Option<ToolCallResponse>> {
-    // Raise not implemented error
-    Err(anyhow::anyhow!("Not implemented"))
-}
+pub mod json_parser;
+pub mod parsers;
+pub mod pythonic_parser;
+pub mod response;
+pub mod tools;
+pub mod xml_parser;
