+++ lib/llm/src/postprocessor/tool_calling/xml_parser.rs
@@ -12,3 +12,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
+pub fn try_tool_call_parse_xml(message: &str) -> anyhow::Result<Option<ToolCallResponse>> {
+    // Raise not implemented error
+    Err(anyhow::anyhow!("Not implemented"))
+}
