+++ lib/llm/src/preprocessor/tools/mod.rs
@@ -13,7 +13,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub fn try_tool_call_parse_pythonic(message: &str) -> anyhow::Result<Option<ToolCallResponse>> {
-    // Raise not implemented error
-    Err(anyhow::anyhow!("Not implemented"))
-}
+pub mod request;
